requirements.txt
streamlit
graphviz
Pillow